% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier_functions.R
\name{bezier_eval}
\alias{bezier_eval}
\title{Find a point on a 2D Bezier curve B(t) at a particular value of t}
\usage{
bezier_eval(p0, p1, p2, t)
}
\arguments{
\item{p0}{Start control point}

\item{p1}{Intermediate control point}

\item{p2}{End control point}

\item{t:}{Location on the curve}
}
\value{
The x-coordinate of B(t)
}
\description{
Find a point on a 2D Bezier curve B(t) at a particular value of t
}
\examples{
bezier_eval(p0 = c(0.4,0.1),p1 = c(0.9,0.9),p2 = c(0.1,0.4),t = 0)

}
