% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_functions.R
\name{eta_analytical}
\alias{eta_analytical}
\title{Evaluate the analytical value of \eqn{\eta} for bivariate Gaussian, logistic, asymmetric logistic, and inverted logistic copulas}
\usage{
eta_analytical(theta, dep = NULL)
}
\arguments{
\item{theta}{Strength of dependence, between (0,1)}

\item{dep}{Copula; \code{g} for Gaussian, \code{l} for logistic, \code{il} for inverted logistic, \code{al} for asymmetric logistic}
}
\value{
scalar value of \eqn{\eta}
}
\description{
Evaluate the analytical value of \eqn{\eta} for bivariate Gaussian, logistic, asymmetric logistic, and inverted logistic copulas
}
