% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_functions.R
\name{eta_empirical}
\alias{eta_empirical}
\title{Compute \eqn{\eta} based on Bezier spline control points, usually a posterior draw obtained from \code{fit_mcmc_bezier}}
\usage{
eta_empirical(control_points)
}
\arguments{
\item{control_points}{A 9x1 vector of Bezier spline control points}
}
\value{
scalar value of \eqn{\eta}
}
\description{
Compute \eqn{\eta} based on Bezier spline control points, usually a posterior draw obtained from \code{fit_mcmc_bezier}
}
