% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_function.R
\name{gen_data_exp}
\alias{gen_data_exp}
\title{Generate data with unit exponential margins from 4 different bivariate copulas}
\usage{
gen_data_exp(n, theta, copula = NULL, tau = 0.75)
}
\arguments{
\item{n}{Number of observations}

\item{theta}{Dependence parameter (between 0 and 1)}

\item{copula}{\code{l} for logistic, \code{il} for inverted-logist, \code{al} for asymmetric logistic, \code{g} for Gaussian}

\item{tau}{Quantile threshold using marginal thresholding}
}
\value{
x n x 2 matrix of data with unit exponential marginal distributions

r Vector of radii

w Vector of angles

r_0_marg truncation point for truncated Gamma distribution for all n data points

above_thresh_marg TRUE/FALSE vector of whether points are above the marginal threshold

data_marg_r_0 List with n, r, w, and r_0 for data points where above_thresh_marg=T
}
\description{
Generate data with unit exponential margins from 4 different bivariate copulas
}
\examples{
simdata <- gen_data_exp(n = 600, theta = 0.8, tau=0.75, copula = 'g')
summary(simdata$x)
}
