% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limit_set_functions.R
\name{get_r0_marg}
\alias{get_r0_marg}
\title{Truncated Gamma calculations}
\usage{
get_r0_marg(x, tau)
}
\arguments{
\item{x}{bivariate data with exponential margins, in a matrix with 2 colums}

\item{tau}{quantile level for marginal threshold}
}
\value{
r_0_marg truncation point for truncated Gamma distribution

above_thresh_marg TRUE/FALSE vector of whether points are above the marginal threshold
}
\description{
Truncated Gamma calculations
}
