% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limit_set_functions.R
\name{gx}
\alias{gx}
\title{Evaluate the rate parameter of the truncated Gamma distribution}
\usage{
gx(N = N, p = P, m = x[, 2]/x[, 1], p_l = x[, 1])
}
\arguments{
\item{N}{Number of data points that the truncated Gamma distribution is to be fitted to}

\item{p}{7x2 matrix of points that define the Bezier spline}

\item{m}{Vector of slopes of length N; entries are slopes of lines connecting the origin to each point}

\item{p_l}{x-coordinate of each point}
}
\value{
Vector of rate parameters of the truncated Gamma distribution
}
\description{
Evaluate the rate parameter of the truncated Gamma distribution
}
\keyword{internal}
