% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_functions.R
\name{lambda_analytical}
\alias{lambda_analytical}
\title{Evaluate the analytical value of \eqn{\lambda(\omega)} for bivariate Gaussian, logistic, asymmetric logistic, and inverted logistic copulas}
\usage{
lambda_analytical(theta, omega, dep = NULL)
}
\arguments{
\item{theta}{Strength of dependence, between (0,1)}

\item{omega}{scalar between (0,1)}

\item{dep}{Copula; \code{g} for Gaussian, \code{l} for logistic, \code{il} for inverted logistic, \code{al} for asymmetric logistic}
}
\value{
scalar value of \eqn{\lambda(\omega)}
}
\description{
Evaluate the analytical value of \eqn{\lambda(\omega)} for bivariate Gaussian, logistic, asymmetric logistic, and inverted logistic copulas
}
