% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_functions.R
\name{lambda_empirical}
\alias{lambda_empirical}
\title{Compute \eqn{\lambda(\omega)} based on Bezier spline control points, usually a posterior draw obtained from \code{fit_mcmc_bezier}}
\usage{
lambda_empirical(control_points, omega)
}
\arguments{
\item{control_points}{A 9x1 vector of Bezier spline control points}

\item{omega}{scalar between (0,1)}
}
\value{
scalar value of \eqn{\lambda(\omega)}
}
\description{
Compute \eqn{\lambda(\omega)} based on Bezier spline control points, usually a posterior draw obtained from \code{fit_mcmc_bezier}
}
