% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limit_set_functions.R
\name{log_lik}
\alias{log_lik}
\title{Evaluate the truncated Gamma likelihood for a vector of radii}
\usage{
log_lik(r, r_0, shape, rates)
}
\arguments{
\item{r}{Vector of radii}

\item{r_0}{Vector of truncation points for each radii}

\item{shape}{Common shape parameter for the truncated Gamma distribution}

\item{rates}{Vector of rate parameters for each truncated Gamma distribution}
}
\value{
Sum of the truncated Gamma log-likelihoods
}
\description{
Evaluate the truncated Gamma likelihood for a vector of radii
}
\examples{
log_lik(r = 10, r_0 = 3, shape = 2, rates = 1)
}
\keyword{internal}
