% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limit_set_functions.R
\name{makeP}
\alias{makeP}
\title{Make a 7x2 matrix that defines a Bezier spline, based on parameters and constraints}
\usage{
makeP(theta, pmix0, pmix1, pmix2, pmix3)
}
\arguments{
\item{theta}{Vector of 9 control points that define the shape of the spline}

\item{pmix0}{Unit mass probability of p0y and p6x}

\item{pmix1}{Unit mass probability for p1x and p5y}

\item{pmix2}{Unit mass probability for p2x and p4y}

\item{pmix3}{Unit mass probability for p3}
}
\value{
P 7x2 matrix that defines a Bezier spline comprised of 3 Bezier curves
}
\description{
Make a 7x2 matrix that defines a Bezier spline, based on parameters and constraints
}
\keyword{internal}
