% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_bezier_polar}
\alias{plot_bezier_polar}
\title{Plot the limit set based on MCMC output in polar coordinates and export median curve}
\usage{
plot_bezier_polar(
  mcmc_samples,
  x,
  r_0_marg,
  above_thresh_marg,
  theta = NULL,
  copula = NULL,
  thin.by = 10,
  plot.fig = T,
  plottitle = "Polar coordinates",
  plot.truth = T
)
}
\arguments{
\item{mcmc_samples}{Output object from mcmc_bezier function}

\item{x}{Original data in exponential margins}

\item{r_0_marg}{Vector of truncation thresholds for all points}

\item{above_thresh_marg}{Vector of TRUE/FALSE values, where TRUE indicates points selected for fitting the truncated Gamma model}

\item{theta}{True value of copula dependence parameter (if known)}

\item{copula}{True parametric copula form (if known)}

\item{thin.by}{Thin the posterior for plotting (and median curve calculation)}

\item{plot.fig}{TRUE/FALSE for whether the limit set should be plotted}

\item{plottitle}{Title of plot (only used if plot.fig = T)}

\item{plot.truth}{TRUE/FALSE for whether the true limit set should be plotted (if known)}
}
\value{
Row of thinned mcmc_samples which corresponds to the median curve
}
\description{
Plot the limit set based on MCMC output in polar coordinates and export median curve
}
