% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_functions.R
\name{tau_analytical}
\alias{tau_analytical}
\title{Evaluate the analytical value of \eqn{\tau_1(\delta)} for bivariate Gaussian, logistic, asymmetric logistic, and inverted logistic copulas}
\usage{
tau_analytical(theta, delta, dep = NULL)
}
\arguments{
\item{theta}{Strength of dependence, between (0,1)}

\item{delta}{scalar between (0,1)}

\item{dep}{Copula; \code{g} for Gaussian, \code{l} for logistic, \code{il} for inverted logistic, \code{al} for asymmetric logistic}
}
\value{
scalar value of \eqn{\tau_1(\delta)}
}
\description{
Evaluate the analytical value of \eqn{\tau_1(\delta)} for bivariate Gaussian, logistic, asymmetric logistic, and inverted logistic copulas
}
