% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_functions.R
\name{tau_empirical}
\alias{tau_empirical}
\title{Compute \eqn{\tau_1(\delta)} based on Bezier spline control points, usually a posterior draw obtained from \code{fit_mcmc_bezier}}
\usage{
tau_empirical(control_points, delta)
}
\arguments{
\item{control_points}{A 9x1 vector of Bezier spline control points}

\item{delta}{scalar between (0,1)}
}
\value{
scalar value of \eqn{\tau_1(\delta)}
}
\description{
Compute \eqn{\tau_1(\delta)} based on Bezier spline control points, usually a posterior draw obtained from \code{fit_mcmc_bezier}
}
