% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limit_set_functions.R
\name{thetaOK}
\alias{thetaOK}
\title{Check whether the parameter vector satisfies the constraints on the limit set}
\usage{
thetaOK(theta, pmix0, pmix1, pmix2, pmix3)
}
\arguments{
\item{theta}{Vector of (9) control points and (unused) truncated Gamma shape parameter}

\item{pmix0}{Unit mass probability of p0y and p6x}

\item{pmix1}{Unit mass probability for p1x and p5y}

\item{pmix2}{Unit mass probability for p2x and p4y}

\item{pmix3}{Unit mass probability for p3}
}
\value{
TRUE/FALSE if constraints are satisfied/not-satisfied
}
\description{
Check whether the parameter vector satisfies the constraints on the limit set
}
\keyword{internal}
